requirement.txt
scrapy
pycrypto
psycopg2
